Fixed score and high score.  Adjusted and added bumpers.
Adjusted the force amounts and positions of lots of bumpers.  Finally got the score and high score systems fixed (thanks Matt).

Started working on a multiply bumper that will instantiate more ball prefabs.

Added a queue that tracks game objects that the ball collides with.  Haven't done anything visual with this yet.