Fixed Sticky Bumper & Multiply Bumper issues by adding a queue to the sticky bumper
Went to Code Help desk this morning and Sienna helped me fix the sticky bumper to make sure it grabs the correct rigidbody, instead of the automatically grabbing the rigidbody of the original ball.

Then we added a queue that keeps track of the rigidbodies that hit the sticky bumper so that they are frozen, unfrozen, and shot out individually and in the same order that the collisions occur.

I'm in the process of adding more comments and deleting old stuff to clean up the scripts a bit.